+++ error/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 authors = ["Danny Meyer <Danny.Meyer@gmail.com>"]
 edition = "2018"
 description = "Errors for WereSoCool"
-license = "MIT"
+license = "GPL-3.0"
 
 [dependencies]
 portaudio = "0.7.0"
