+++ parser/Cargo.toml
@@ -18,5 +18,3 @@ num-rational = "0.2"
 indexmap = "1.3.1"
 weresocool_ast = { path = "../ast" }
 weresocool_error = { path = "../error" }
-#weresocool_ast = "0.5.0"
-#weresocool_error = "0.1.0"
