+++ src/testing/expect_tests/mocks/index_with_term.socool
