+++ src/testing/expect_tests/mocks/list_composed_with_op.socool
