+++ src/testing/expect_tests/mocks/list_fit_length.socool
