+++ src/testing/expect_tests/mocks/list_indexed_fit_length.socool
