+++ src/testing/expect_tests/mocks/list_named_indexed_fit_length.socool
