+++ src/testing/expect_tests/mocks/list_named_with_indicies.socool
