+++ src/testing/expect_tests/mocks/list_with_random.socool
