+++ src/testing/expect_tests/mocks/simple.socool
