+++ src/testing/expect_tests/mocks/simple_fun.socool
