+++ src/testing/mod.rs
@@ -15,8 +15,7 @@ use std::hash::{Hash, Hasher};
 use std::io::Write;
 use weresocool_ast::{NormalForm, Normalize as NormalizeOp};
 use weresocool_parser::*;
-pub mod expect;
-mod tests;
+mod expect_tests;
 
 type TestTable = IndexMap<String, CompositionHashes>;
